backend/requirements.txt
attrs==19.3.0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
more-itertools==8.1.0
packaging==20.0
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytesseract==0.3.1
pytest==5.3.2
python-dotenv==0.10.3
six==1.13.0
wcwidth==0.1.8
Werkzeug==0.16.0
zipp==0.6.0
